feat(notification_init): 알림 생성 응답에서 알림개수 필드 long -> int로 변경

+++ src/main/java/me/jinjjahalgae/domain/notification/dto/NotificationCreateResponse.java
@@ -6,7 +6,7 @@ package me.jinjjahalgae.domain.notification.dto;
  * @param message 생성된 메세지 확인용
  */
 public record NotificationCreateResponse(
-        Long notificationCnt,
+        int notificationCnt,
         String message
 ) {
 }
